tests/wpt/web-platform-tests/css/css-text-decor/text-decoration-thickness-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Decoration Test: parsing text-decoration-thickness longhands with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-decor-4/#text-decoration-width-property">
<meta name="assert" content="text-decoration-thickness supports the following values: auto | from-font | <length>">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("text-decoration-thickness", "auto");
test_valid_value("text-decoration-thickness", "from-font");
test_valid_value("text-decoration-thickness", "-10px");
test_valid_value("text-decoration-thickness", "2001em");
test_valid_value("text-decoration-thickness", "-49em");
test_valid_value("text-decoration-thickness", "53px");
test_valid_value("text-decoration-thickness", "calc(40em - 10px)");
test_valid_value("text-decoration-thickness", "calc(-50em + 13px)");
test_valid_value("text-decoration-thickness", "10%");
test_valid_value("text-decoration-thickness", "-27%");
test_valid_value("text-decoration-thickness", "calc(40% - 20px)");
test_valid_value("text-decoration-thickness", "calc(100% - 40em)");
</script>
</body>
</html>
